requirements.txt
certifi==2019.3.9
Click==7.0
colorama==0.4.1
Flask==1.0.2
Flask-Assets==0.12
Flask-Cors==3.0.7
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
peewee==3.9.2
peewee-db-evolve==3.7.0
psycopg2-binary==2.7.7
PyJWT==1.7.1
python-dotenv==0.10.1
six==1.12.0
webassets==0.12.1
Werkzeug==0.15.1
